applications/batch-inference/requirements.txt
tensorflow-serving-api==1.9.0
